profile user
upload profile,

+++ APP_NEW/app/src/main/java/com/example/vuivcfunnyapp/ui/profile/profile_interface/TruyenDuLieu.java
@@ -1,7 +1,7 @@
 package com.example.vuivcfunnyapp.ui.profile.profile_interface;
 
 public interface TruyenDuLieu {
-    public void GetSoluongVideo(int soluong);
-    public void GetSoluongFollower(int soluong);
-    public void GetSoluongFollowing(int soluong);
+    public void GetSoluongVideo(long soluong);
+    public void GetSoluongFollower(long soluong);
+    public void GetSoluongFollowing(long soluong);
 }
